functions/add-user/requirements.txt
functions-framework==3.*
google-cloud-firestore==2.*
pytz==2024.1
phonenumbers==8.13.37
